finished implementing most of level1-1 objects
The objects are place precisely where they are in the original NES game
however scaled to the 16:9 ratio. Also added SolidBlock which just is
NormalBlock with removal of popUp when mario collides upwards with it.
and the use of stairs functions to create stairs and stacks of blocks
helped with declaring specific positions for blocks while minimising
code.