text class and preLevel instead of levelSelect
I added a text class so that text is now an object and can have it’s
own width, x and y values this allows me to use the multiplier function
i created before so when the screen resizes it automatically resizes
the text to the screen. I also changed the levelSelect function to the
preLevel function to further resemble the mario NES that i am basing
this game off.