gameOver added, time added and other minor additions
The gameOver function is a method of mario class which resets the game
while depreciating a life and showing the user he has died, taking him
back to the preLevel scene. The timing in the corner is just like the
normal game however i think the normal game has a slightly distorted
timing system where it starts from 400 and counts down ever 1/2 or 3/4
of a second. I have put the reseting of the level values to when the
user clicks new game on main scene. This makes the user only be able to
see his most previous score. I was thinking about making the variable
local however ran into some issues so didn’t bother.